Update and rename Talent_Scraper.command to Talents_Fetch_HoTS_Mac.command
Don’t know which talents to pick? 
Is the Icy Veins website too distracting?
Launch this file and simply type the name of the hero which you will be, and it will show all the talent sets given on the Icy Veins website in a simple “Level : Talent Name” format, so you can keep your focus on the game and less on the loud website interface.